pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opindex"
version = "0.1.0"
description = "Linux-first command snippet manager for markdown notes"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Cameron Candau" }
]
dependencies = [
  "PyYAML>=6.0",
]
keywords = ["cli", "rofi", "clipboard", "markdown", "notes", "commands"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Environment :: X11 Applications",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/CameronCandau/OpIndex"
Repository = "https://github.com/CameronCandau/OpIndex"
Issues = "https://github.com/CameronCandau/OpIndex/issues"

[project.scripts]
opindex = "cmd_manager.cli:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["cmd_manager*"]
